* Fixed an issue with GRAPHICAL wildcards causing problems when using Ninja. Fixed a vector constructor which MSVC really dislikes. Also had to simplify headless_triangle_test.cpp because MSVC gets an internal compiler error when trying to build it (after fixing the dodgy vector ctor). Pretty sure the code was all correct and MSVC is just bad

+++ src/gl/impl/ogl/renderer.cpp
@@ -355,7 +355,7 @@ namespace tz::gl
     {
         GLfloat rgba[4];
         glGetFloatv(GL_COLOR_CLEAR_VALUE, rgba);
-        return {rgba[0], rgba[1], rgba[2], rgba[3]};
+        return {static_cast<float>(rgba[0]), static_cast<float>(rgba[1]), static_cast<float>(rgba[2]), static_cast<float>(rgba[3])};
     }
 
     void RendererOGL::render()
